Add paging support, and run the kernel in upper memory.
This is a necessary step for supporting a userland with memory placement
restrictions. Often, ELF binaries will want to be placed at the fixed location
where the kernel is also placed. This is only possible with virtual memory, and
only if the kernel is mapped elsewhere. This commit does not move the kernel in
physical memory, but it enables virtual memory with a mapping with place for
new processes.

Note that after this commit, as soon as a userland process starts running, a
page fault occurs; this is because the current "userland" actually runs code
from kernel space. However, as of this commit this code is inaccessible from
the userland because it is mapped kernel-only. The access causes a "page
protection violation during page read in unprivileged mode". This is expected
and will be fixed by having actual userland processes with their own code
pages.