Add elfrun, an UDP receiver for ELF binaries that will run them as userland processes.
Also raise the initially available frames in the virtio implementation a bit,
since we don't seem to recycle them well currently and this allows us to run a
few of these processes.

The elfrun receiver sends a UDP packet to some destination set in
kernel_main.cpp, with a filename in payload. I wrote a small Perl script that
responds to such UDP packets by sending back a file with that name, in this
case, the one built in userland/ since this commit. The elfrun implementation
will receive that binary and run it as a userland process.