Add file descriptor table to a process and implement putstring() in terms of it.
A process now always starts with file descriptor 0, which is the VGA stream to
the console. The syscall putstring() has a new first parameter with the fd
index (so this parameter must be 0 at this moment). If the fd and string
addresses seem valid, fd->putstring() is called, which writes the message on
the screen.