Implement sys_mem_map() and sys_mem_unmap(). Pass fd 0 to exec_test in the right way.
Note that there is a bug in cloudlibc pre-0.51 that causes jemalloc not to work
on i686. If cloudlibc 0.51 (or post-0.50 HEAD) is used, after this commit,
malloc() can be used to allocate memory.

This also means that argdata_create_fd() can be used to pass fd's, so we can
remove the VGA stream hack. Also, I fixed a little bug in the passing of fd 0.