Add a condition variable with wait(), notify() and notify_all(), that can already block and unblock threads in the scheduler.
When a thread is blocked, it is not immediately removed from the ready list in
the scheduler. Instead, when the scheduler sees a blocked thread in the ready
list, it unschedules it then, and notes this in the thread. Then, when the
thread is ready again, it is marked non-blocked, and rescheduled if it was
unscheduled. This saves walking through the ready-list when a thread is
blocked, and causes no actual changes to the ready-list if a thread is
unblocked before the scheduler even sees it was blocked.