Add pipe_fd, an unnamed pipe implementation, including a userland test for it.
The pipe_fd is the first user of the cv_t, which means that if you read() on it
while there is no data available, your thread is temporarily blocked and will
not be scheduled until someone calls write() on it. This will notify the
condition variable guarding the "not empty" condition, so all threads waiting
get a chance to take some data from it again.