Allow giving a path to the CloudABI unittest binary, which init will then run.
Since this binary is 3 MB on my machine, the kernel grew to more than 4 MB, so
identity mapping only the first huge page was not enough anymore. While the
kernel eventually identity maps everything it needs, the bootstrapping code
only maps the first huge page --- and since the initial bootstrap stack was
placed after the code by my linker, the first usage of the stack in upper
memory caused a fault. So for now, I identity map the first 12 MB while
bootstrapping.

Also, the vDSO has a bug in some places where not all output operands are
marked as such, causing some code to fail to return an error correctly. I've
corrected this for all functions as well in this commit.