Beginnings of a thread_condition framework, for blocking threads on multiple things simultaneously (e.g. poll()).
The framework consists of three major structs:
- the thread condition, "something a thread wants to wait for"
- the thread condition signaler, "something which satisfies a condition and
  unblocks a thread"
- the thread condition waiter, "something which blocks on behalf of a thread"

After this commit, the framework is used only in cv_t::wait() for a CV to block
on itself, but I'll soon implement it in the fd's and locks as well, so I can
implement poll() with it. Primarily, as soon as proc_terminate is a waitable
thread condition, pdwait() already works so a lot of new CloudABI unittests
will start to succeed.