Before blocking ourselves on a thread_condition, check if it was already satisfied.
The thread_condition_waiter asks the thread_condition_signaler if the
thread_condition is already satisfied. For this, the thread_condition_signaler
gets a function pointer from its owner, which returns whether the condition is
satisfied. For example, the process_fd supplies its termination signaler with a
function pointer that returns whether the process has exited.

If the thread_condition_waiter notices that one of its conditions is already
initially-satisfied, it doesn't need to wait for more conditions. It skips the
blocking of its threads and immediately cancels all subscriptions it has done
until now, so such a wait() returns immediately.