In the scheduler, halt the CPU when no thread is ready to run.
In this situation, the scheduler will mark itself as 'waiting for ready task',
re-enable interrupts and halt the processor. The next interrupt will wake up
the processor, and will be handled as normal. If it is a timer interrupt,
however, the interrupt handler will see that the scheduler is already waiting
for ready task, and will simply iret. Then, execution of the scheduler will
continue as normal; if a thread became ready (because of new hardware state or
simply because of time passing) it will now be scheduled, otherwise the
processor will be halted again.