Fix an occasional userland crash because of returning to an exited thread after sys_proc_exec(). Also add assert().
While sys_proc_exec() should only return on error, accidentally it could also
return success. In debug mode, this never caused a crash, possibly because the
timer would interrupt the thread, which was already unscheduled so wouldn't be
scheduled anymore after being interrupted. In release mode, however, it was
possible that the timer didn't interrupt the thread, so it had enough time to
return from sys_proc_exec(); however, since the stack was reinitialized, this
caused a return to address 0x0, causing a page fault. This was fixed by having
a thread yield after a successful sys_proc_exec().

Also add an implementation of assert(), and two assertions to verify that this
bug was correctly diagnosed and now fixed.