Fix 'out of memory' error in release mode by correctly initializing the page directory.
Because the page directory wasn't initialized, its allocation could be filled
with data from a previous allocation. This could cause the userland page mapper
to assume a physical page was already allocated for a virtual one, with an
uninitialized physical address. Then, when the virtual page was deallocated,
the uninitialized physical address was returned to the 'free physical page'
pool. In this case, it turned out that the uninitialized physical address was
quite low, so by re-aligning it it became 0. Therefore, the address 0 was
re-added to the free physical page pool. The next allocation from the physical
page pool therefore returned 0, causing the out-of-memory error while plenty
memory was still available.