Fixed killing a process with subchilds, and added a test for it.
From now on, exiting threads may no longer call thread_yield(), as it would
cause the thread to no longer be scheduled and its resources not to be cleaned
up. Instead, such threads unroll all the way back to the interrupt handler,
which calls thread_final_yield() to unschedule and clean up the thread.