Implement accept() on UDP sockets, to be able to respond to listening UDP sockets in CloudABI.
Since cloudabi_recv_out_t contains a cloudabi_sockaddr_t member, it is possible
to implement recvfrom() in Cloudlibc. A CloudABI binary that is started with a
listening (bound but unconnected) UDP socket, could use this method to obtain
messages and their senders. However, there is no way to sendto() a message back
to them, by design, as sending UDP packets to blanket destinations would break
the "injected capabilities" model of CloudABI.

However, listening UDP sockets and responding to incoming messages is an
important use-case for a lot of protocols like DNS, NTP, NFS or BitTorrent.
Therefore, I discussed this with Ed Schouten, and he offered the idea to
implement accept() for UDP sockets. If messages are waiting on the socket, a
new FD would be created that is connected to the origin of the first message,
and the message would be moved to it. After this, the message can be read from
the socket, and it can be used to respond as well. Since the new socket is also
bound to the destination of its received packet, this also solves the issue of
IP source address selection.

This commit implements this approach. It also implements socket-type pseudo FDs
up to the point that accept, stat_get, recv and send work in libpseudo,
exposing accept() to the userland. It also corrects some related issues in the
networkd.