Add an early but mostly functional TCP socket implementation.
Both listen and connecting sockets work. Receiving out-of-band data is not
supported; such segments will be ignored causing retransmission on the peer
side. This implementation does not retransmit lost packets, it does not
implement timeout timers, it does not understand triple-ACK or other kinds of
congestion or loss signals, and it uses a fixed-size window. So there is lots
of room for improvement.