When the user types, send the keystrokes to the console, instead of immediately outputting.
From there, the keystrokes will be interpreted and eventually become readable
by the userland as stdin. If echoing is enabled (the default), non-escape codes
will also be shown on the screen. You can already type the Cosix-specific
escape code <ESC>[?8001h, and echoing will be turned on. However, since then
characters don't appear on the screen anymore, typing <ESC>[?8002h will not
turn echoing back on; an application will have to write that.