Fix a bug that occurred when thread_condition_waiter::wait() was called recursively.
This could happen if the already_satisfied_function of one of the waiter's
conditions itself blocked the thread. For example, the pseudo FD implementation
did this. In this case, we have an outer wait() and an inner wait(); it is
possible that the outer wait() conditions would wake up the thread during the
inner wait() block. In that case, the inner waiter would become confused
because none of its own conditions was satisfied, yet it was unblocked.

This commit ensures that no condition links to the active thread during a call
to an already satisfied function, so this situation cannot occur. It also adds
an extra check for this situation.