Implement timed lock/condvar polling.
The thread functions for lock/condvar waiting don't themselves block
anymore; they return signalers which can be used in the normal thread
condition waiter mechanism in poll. On the other hand, now that waiting
for a lock or condvar can be interrupted, cancelling of this waiting
has been added as well.