Implement support for O_NONBLOCK in all places where fds would currently block.
The one exception is the pseudo FD: here, transmission of a reverse request
must wait if another request is in-flight, because it needs to be answered
first. This is a design flaw; a blocking read() that will never return will
cause other nonblocking calls to have to wait. A redesign of the reverse
protocol to allow multiple in-flight requests will be necessary to solve this.