Add an ext2 filesystem driver.
Starting from this commit, the system will read the cmdline and get a specification
for the config filesystem. The specification can be 'tmpfs@', meaning a temporary
filesystem, or 'ext2@ata0p0', meaning an ext2 filesystem on the first partition of
the first ATA device.

Currently, this means that the config filesystem will be mounted, and the userland
tests will be run on top of it. In the default CMake 'boot' target, the value is
'ext2@ata0p0'.